+++ crates/bevy_transform/Cargo.toml
@@ -15,7 +15,7 @@ bevy_ecs = { path = "../bevy_ecs", version = "0.14.0-dev", features = [
   "bevy_reflect",
 ], optional = true }
 bevy_hierarchy = { path = "../bevy_hierarchy", version = "0.14.0-dev", optional = true }
-bevy_math = { path = "../bevy_math", version = "0.14.0-dev" }
+bevy_math = { path = "../bevy_math", version = "0.14.0-dev", default-features = false }
 bevy_reflect = { path = "../bevy_reflect", version = "0.14.0-dev", features = [
   "bevy",
 ], optional = true }
@@ -24,21 +24,28 @@ thiserror = "1.0"
 
 [dev-dependencies]
 bevy_tasks = { path = "../bevy_tasks", version = "0.14.0-dev" }
-bevy_math = { path = "../bevy_math", version = "0.14.0-dev", features = [
+bevy_math = { path = "../bevy_math", version = "0.14.0-dev", default-features = false, features = [
   "approx",
 ] }
 approx = "0.5.1"
 
 [features]
-bevy_app = ["dep:bevy_app"]
-bevy_ecs = ["dep:bevy_ecs"]
-bevy_hierarchy = ["dep:bevy_hierarchy"]
-bevy_reflect = ["dep:bevy_reflect"]
+# Adds normal Bevy impls like deriving components, bundles, reflection, as well as adding
+# systems for transform propagation and more.
+# This exists because it allows opting out of all of this, leaving only a bare-bones transform struct,
+# which enables users to depend on that without needing the larger Bevy dependency tree.
+bevy_impls = [
+  "dep:bevy_app",
+  "dep:bevy_ecs",
+  "dep:bevy_hierarchy",
+  "dep:bevy_reflect",
+  "bevy_math/bevy_reflect",
+]
+
+default = ["bevy_impls"]
 
 # Turning off default features leaves you with a barebones
 # definition of transform.
-default = ["bevy_app", "bevy_ecs", "bevy_hierarchy", "bevy_reflect"]
-
 serialize = ["dep:serde", "bevy_math/serialize"]
 
 [lints]
