+++ crates/bevy_transform/src/components/global_transform.rs
@@ -2,11 +2,9 @@ use std::ops::Mul;
 
 use super::Transform;
 
-#[cfg(feature = "bevy_ecs")]
-use bevy_ecs::{component::Component, reflect::ReflectComponent};
+use bevy_ecs::component::Component;
 use bevy_math::{Affine3A, Dir3, Mat4, Quat, Vec3, Vec3A};
-#[cfg(feature = "bevy_reflect")]
-use bevy_reflect::{std_traits::ReflectDefault, Reflect};
+use bevy_reflect::Reflect;
 
 /// Describe the position of an entity relative to the reference frame.
 ///
@@ -36,11 +34,8 @@ use bevy_reflect::{std_traits::ReflectDefault, Reflect};
 /// - [`transform`]
 ///
 /// [`transform`]: https://github.com/bevyengine/bevy/blob/latest/examples/transforms/transform.rs
-#[derive(Debug, PartialEq, Clone, Copy)]
-#[cfg_attr(feature = "bevy_ecs", derive(Component))]
+#[derive(Component, Debug, PartialEq, Clone, Copy, Reflect)]
 #[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
-#[cfg_attr(feature = "bevy_reflect", derive(Reflect))]
-#[cfg_attr(feature = "bevy_reflect", reflect(Component, Default, PartialEq))]
 pub struct GlobalTransform(Affine3A);
 
 macro_rules! impl_local_axis {
