+++ crates/bevy_transform/src/components/mod.rs
@@ -1,5 +1,7 @@
+#[cfg(feature = "bevy_impls")]
 mod global_transform;
 mod transform;
 
+#[cfg(feature = "bevy_impls")]
 pub use global_transform::*;
 pub use transform::*;
